Prevent inlining of Logger.Create()
Logger.Create() depends on analyzing the call-stack to determine who it
is creating a Logger for.

In the -current- workspacer-codebase this probably works fine.

But in my experience, especially in optimized release-builds, some
method-calls tend to get inlined, making this kind of analyzis
unreliable, and worst case causes bugs or misleading logs.

We can prevent this from ever happening by instructing the compiler to
never inline this method-call.

So let's do just that :)