Reduce nesting in Workspace.cs. Improve clarity of if-branches.
This class has some big, nested if-branches where there are always 1
short alternate and big one.

Making the short follow the big one makes it hard to position it in
the if-tree, and makes the code harder to reason about.

Invert the 2 ifs, and promote the smaller-usecases to the top to make
total if-statement much easier to reason about.