Remember current workspace before switching
This commit adds a `_lastWorkspace` variable to WorkspaceManager. This
variable is set to the currently focused workspace just prior to
switching to a different workspace.

This allows implementation of a `SwitchToLastFocusedWorkspace()` method
that will switch to the workspace stored in `_lastWorkspace`. This
allows switching back and forth between the currently focused workspace
and the previously focused workspace on each call.

A default keybind is added to call this method via `mod+tilde`.