Added a beta release branch (#227)
This gives us the ability to have beta releases.

This is based on the `unstable` workflow.

To create a new release:

1. Create `beta` tag on your local branch
2. Push to the remote

The workflow builds workspacer and generates an installer. It deletes the `beta` tag and replaces it with the version number.

The version number is formatted like `v0.9.11-beta.2`. The final number in the versioning relies on the `github.run_number` - this means that it's possible to have `v0.9.11-beta.2` followed by `v0.9.12-beta.3`.

Co-authored-by: Jostein Kjønigsen <jostein@kjonigsen.net>